requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.42.0
django-autoslug==1.9.7
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-embed-video==1.3.3
django-filter==2.2.0
django-tinymce==3.0.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
h5py==2.9.0
idna==2.9
importlib-metadata==1.6.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.2.2
numpy==1.16.2
oauthlib==3.1.0
Pillow==7.1.2
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.1
requests==2.23.0
requests-oauthlib==1.3.0
scipy==1.2.1
six==1.12.0
sqlparse==0.3.1
urllib3==1.25.9
zipp==3.1.0
